+++ CHANGELOG.md
@@ -1,5 +1,38 @@
 # oslo
 
+## 1.1.3
+
+### Patch changes
+
+- Update dependencies. ([#51](https://github.com/pilcrowOnPaper/oslo/pull/51))
+
+## 1.1.2
+
+### Patch changes
+
+- Export OAuth2 token response interface ([#45](https://github.com/pilcrowOnPaper/oslo/pull/45))
+- Fix client data JSON validation
+
+## 1.1.1
+
+### Patch changes
+
+- Improve `base64.encode()` performance
+- Improve `base32.encode()` performance. ([#40](https://github.com/pilcrowOnPaper/oslo/pull/40))
+
+## 1.1.0
+
+### Minor changes
+
+- Deprecate `encodeBase32()`, `decodeBase32()`, `encodeBase64`, `decodeBase64()`, `encodeBase64url()`, `decodeBase64url()`. ([#35](https://github.com/pilcrowOnPaper/oslo/pull/35))
+- Feat: Add `Base64Encoding`, `Base32Encoding`, `base16`, `base32`, `base32hex`, `base64`, `base64url`. ([#35](https://github.com/pilcrowOnPaper/oslo/pull/35))
+
+## 1.0.4
+
+### Patch changes
+
+- Fix: Export `TimeSpanUnit` ([#30](https://github.com/pilcrowOnPaper/oslo/pull/30))
+
 ## 1.0.3
 
 - Pin `@node-rs/argon2` and `@node-rs/bcrypt` versions [#26](https://github.com/pilcrowOnPaper/oslo/pull/26)